added genes vs umis plot
